Upgrade global 'aws-sdk' package
We need to do this temporarily because if we try to match the version that Amplify uses, we encounter problems refreshing the authentication credentials. 

We'll try to unify the versions again once '@aws-amplify/core' updates its dependency to a more up-to-date version of the SDK that does not present this problem.